# torque-api-spec.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Torque API Reference",
    "description": "This page contains information about the Torque APIs and how to use it.<br><br>To be able to test the API methods, you will need an access token to be set in the Authorize section.If you got to this page from your Torque account, we already made it available for you, but you can switch to a different token as needed.<br>To get a different token, use the 'Get token' function available under the 'Access Tokens' section. After running the method, copy the access_token you get in the response and set it in the Authorize fieldas 'Bearer access_token'. For example: Bearer fqSWuw72BbUVFn8AbokF77GJ0r5KEn9MiZjLXF8kBwI.<br><br>The API can be accessed with your account endpoint as will be mentioned in the descriptions and examples or using https://qtorque.io as the prefix instead."
  },
  "paths": {
...
    "/api/spaces/{space_name}/production": {
      "get": {
        "tags": [
          "Production Environments"
        ],
        "summary": "Get production environments",
        "description": "Gets all production environments from a space.",
        "parameters": [
          {
            "name": "space_name",
            "in": "path",
            "description": "Name of the space containing the desired production environments",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductionResponseListItem"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Operation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Space not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Production Environments"
        ],
        "summary": "Create blue environment",
        "description": "Creates a blue production environment.",
        "parameters": [
          {
            "name": "space_name",
            "in": "path",
            "description": "Name of the space that will contain the blue production environment",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductionBlueRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Operation started",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateEnvResponse"
                }
              }
            }
          },
          "400": {
            "description": "Operation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Space or user not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Could not process the request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "424": {
            "description": "Cloud account not accessible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/spaces/{space_name}/production/{production_id}": {
      "get": {
        "tags": [
          "Production Environments"
        ],
        "summary": "Get blue environment",
        "description": "Gets a blue production environment.",
        "parameters": [
          {
            "name": "space_name",
            "in": "path",
            "description": "Name of the space containing the blue production environment",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "production_id",
            "in": "path",
            "description": "Production environment ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductionBlueResponse"
                }
              }
            }
          },
          "400": {
            "description": "Operation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Space or production environment not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Production Environments"
        ],
        "summary": "Delete blue environment",
        "description": "Deletes a blue production environment.",
        "parameters": [
          {
            "name": "space_name",
            "in": "path",
            "description": "Name of the space containing the blue production environment",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "production_id",
            "in": "path",
            "description": "Production environment ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Operation started"
          },
          "400": {
            "description": "Operation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Space or production environment not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "423": {
            "description": "Production environment temporarily unavailable for termination",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "424": {
            "description": "Cloud account not accessible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/spaces/{space_name}/production/{production_id}/debuggingservice": {
      "put": {
        "tags": [
          "Production Environments"
        ],
        "summary": "Update blue environment debugging service",
        "description": "Updates debugging service of blue production environment.",
        "parameters": [
          {
            "name": "space_name",
            "in": "path",
            "description": "Name of the space containing the blue production environment",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "production_id",
            "in": "path",
            "description": "Production environment ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "value",
            "in": "query",
            "description": "New value of debugging service: \"on\" or \"off\"",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "202": {
            "description": "Operation started"
          },
          "400": {
            "description": "Operation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Space, production environment, or QualiY not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Current status of QualiY conflicts with new value",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Value is empty or invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "424": {
            "description": "Cloud account not accessible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/spaces/{space_name}/production/{production_id}/green/debuggingservice": {
      "put": {
        "tags": [
          "Production Environments"
        ],
        "summary": "Update green environment debugging service",
        "description": "Updates debugging service of green production environment.",
        "parameters": [
          {
            "name": "space_name",
            "in": "path",
            "description": "Name of the space containing the green production environment",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "production_id",
            "in": "path",
            "description": "Production environment ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "value",
            "in": "query",
            "description": "New value of debugging service: \"on\" or \"off\"",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "202": {
            "description": "Operation started"
          },
          "400": {
            "description": "Operation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Space, blue/green production environment, or QualiY not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Current status of QualiY conflicts with new value",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Value is empty or invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "424": {
            "description": "Cloud account not accessible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/spaces/{space_name}/production/{production_id}/green/promote": {
      "put": {
        "tags": [
          "Production Environments"
        ],
        "summary": "Promote green environment",
        "description": "User traffic will be redirected only to green production environment.\r\nSee https://community.qtorque.io/creating-your-production-environment-56/managing-production-with-blue-green-deployment-231",
        "parameters": [
          {
            "name": "space_name",
            "in": "path",
            "description": "Name of the space containing the green production environment",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "production_id",
            "in": "path",
            "description": "Production environment ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "202": {
            "description": "Operation started"
          },
          "400": {
            "description": "Operation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Space, blue/green production environment, or account not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Green not ready for promoting or Green cannot be promoting while being exposed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "424": {
            "description": "Cloud account not accessible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateEnvResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Environment ID",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Create environment response",
        "x-internal": true
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Error message",
            "nullable": true
          },
          "name": {
            "type": "string",
            "description": "Error name",
            "nullable": true
          },
          "code": {
            "type": "string",
            "description": "Error code",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Error information",
        "x-internal": true
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Error"
            },
            "description": "Array of errors",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Errors received from an API",
        "x-internal": true
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\"",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}